// src/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          operationName?: string
          query?: string
          variables?: Json
          extensions?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      courses: {
        Row: {
          ay: number
          course_no: string
          id: number
          instructor: string | null
          lang: string | null
          modified_at: string
          rgno: number
          room: string | null
          schedule: Json | null
          season: string
          title_e: string
          title_j: string
          unit: string | null
        }
        Insert: {
          ay: number
          course_no: string
          id?: number
          instructor?: string | null
          lang?: string | null
          modified_at?: string
          rgno: number
          room?: string | null
          schedule?: Json | null
          season: string
          title_e: string
          title_j: string
          unit?: string | null
        }
        Update: {
          ay?: number
          course_no?: string
          id?: number
          instructor?: string | null
          lang?: string | null
          modified_at?: string
          rgno?: number
          room?: string | null
          schedule?: Json | null
          season?: string
          title_e?: string
          title_j?: string
          unit?: string | null
        }
        Relationships: []
      }
      registrations: {
        Row: {
          ay: number
          color: string | null
          id: number
          inserted_at: string
          is_enrolled: boolean
          note: string | null
          rgno: number
          season: string
          user_id: string
        }
        Insert: {
          ay: number
          color?: string | null
          id?: number
          inserted_at?: string
          is_enrolled?: boolean
          note?: string | null
          rgno: number
          season: string
          user_id: string
        }
        Update: {
          ay?: number
          color?: string | null
          id?: number
          inserted_at?: string
          is_enrolled?: boolean
          note?: string | null
          rgno?: number
          season?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "fk_courses"
            columns: ["ay", "season", "rgno"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["ay", "season", "rgno"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof PublicSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof PublicSchema["CompositeTypes"]
    ? PublicSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never